requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
google-cloud-storage